Crate matrix_sdk_crypto_js

matrix-sdk-crypto-js

Welcome to the WebAssembly + JavaScript binding for the Rust matrix-sdk-crypto library! WebAssembly can run anywhere, but these bindings are designed to run on a JavaScript host. These bindings are part of the matrix-rust-sdk project, which is a library implementation of a Matrix client-server.

matrix-sdk-crypto is a no-network-IO implementation of a state machine, named OlmMachine, that handles E2EE (End-to-End Encryption) for Matrix clients.

Usage

These WebAssembly bindings are written in Rust. To build them, you need to install the Rust compiler, see the Install Rust Page. Then, the workflow is pretty classical by using yarn, see the Downloading and installing Node.js and npm Page and installing yarn.

Once the Rust compiler, Node.js and yarn are installed, you can run the following commands:

$ yarn install
$ yarn build
$ yarn test

A matrix_sdk_crypto.js, matrix_sdk_crypto.d.ts and a matrix_sdk_crypto_bg.wasm files should be generated in the pkg/ directory.

TBD

Documentation

The documentation can be found online.

To generate the documentation locally, please run the following command:

$ yarn doc

The documentation is generated in the ./docs directory.

Modules

attachment

Attachment API.

device

Types for a Device.

encryption

Encryption types & siblings.

events

Types related to events.

identifiers

Types for Matrix identifiers for devices, events, keys, rooms, servers, users and URIs.

identities

User identities.

machine

The crypto specific Olm objects.

olm

Olm types.

requests

Types to handle requests.

responses

Types related to responses.

store

Store types.

sync_events

GET /_matrix/client/*/sync

types

Extra types, like Signatures.

verification

Different verification types.

vodozemac

Vodozemac types.

Macros

impl_from_to_inner

We have the following pattern quite often in our code:

Structs

Versions

Object containing the versions of the Rust libraries we are using.

Functions

get_versions

Get the versions of the Rust libraries we are using.

start

Run some stuff when the Wasm module is instantiated.