Struct matrix_sdk_crypto_js::device::Device

pub struct Device { /* private fields */ }

A device represents a E2EE capable client of an user.

Implementations

impl Device

pub fn request_verification( &self, methods: Option<Array> ) -> Result<Promise, JsError>

Request an interactive verification with this device.

Returns a Promise for a 2-element array [VerificationRequest, ToDeviceRequest].

pub fn is_verified(&self) -> bool

Is this device considered to be verified.

This method returns true if either the is_locally_trusted method returns true or if the is_cross_signing_trusted method returns true.

pub fn is_cross_signing_trusted(&self) -> bool

Is this device considered to be verified using cross signing.

pub fn is_cross_signed_by_owner(&self) -> bool

Is this device cross-signed by its owner?

pub fn set_local_trust(&self, local_state: LocalTrust) -> Promise

Set the local trust state of the device to the given state.

This won’t affect any cross signing trust state, this only sets a flag marking to have the given trust state.

trust_state represents the new trust state that should be set for the device.

pub fn user_id(&self) -> UserId

The user ID of the device owner.

pub fn device_id(&self) -> DeviceId

The unique ID of the device.

pub fn display_name(&self) -> Option<String>

Get the human readable name of the device.

pub fn get_key( &self, algorithm: DeviceKeyAlgorithmName ) -> Result<Option<DeviceKey>, JsError>

Get the key of the given key algorithm belonging to this device.

pub fn curve25519_key(&self) -> Option<Curve25519PublicKey>

Get the Curve25519 key of the given device.

pub fn ed25519_key(&self) -> Option<Ed25519PublicKey>

Get the Ed25519 key of the given device.

pub fn keys(&self) -> Map

Get a map containing all the device keys.

pub fn algorithms(&self) -> Array

Get the list of algorithms this device supports.

Returns Array<EncryptionAlgorithm>.

pub fn signatures(&self) -> Signatures

Get a map containing all the device signatures.

pub fn local_trust_state(&self) -> LocalTrust

Get the trust state of the device.

pub fn is_locally_trusted(&self) -> bool

Is the device locally marked as trusted?

pub fn is_blacklisted(&self) -> bool

Is the device locally marked as blacklisted?

Blacklisted devices won’t receive any group sessions.

pub fn is_deleted(&self) -> bool

Is the device deleted?

pub fn first_time_seen(&self) -> u64

Timestamp representing the first time this device has been seen (in milliseconds).