Struct matrix_sdk_crypto_js::machine::OlmMachine

pub struct OlmMachine { /* private fields */ }

State machine implementation of the Olm/Megolm encryption protocol used for Matrix end to end encryption.

Implementations

impl OlmMachine

pub fn new() -> Result<OlmMachine, JsError>

Constructor will always fail. To create a new OlmMachine, please use the initialize method.

Why this pattern? initialize returns a Promise. Returning a

pub fn initialize( user_id: &UserId, device_id: &DeviceId, store_name: Option<String>, store_passphrase: Option<String> ) -> Promise

Create a new OlmMachine.

The created machine will keep the encryption keys either in a IndexedDB based store, or in a memory store and once the objects is dropped, the keys will be lost.

Arguments

user_id - represents the unique ID of the user that owns this machine.

device_id - represents the unique ID of the device that owns this machine.

store_name - The name that should be used to open the IndexedDB based database. If this isn’t provided, a memory-only store will be used. Note the memory-only store will lose your E2EE keys when the OlmMachine gets dropped.

store_passphrase - The passphrase that should be used to encrypt the IndexedDB based

pub fn user_id(&self) -> UserId

The unique user ID that owns this OlmMachine instance.

pub fn device_id(&self) -> DeviceId

The unique device ID that identifies this OlmMachine.

pub fn identity_keys(&self) -> IdentityKeys

Get the public parts of our Olm identity keys.

pub fn display_name(&self) -> Promise

Get the display name of our own device.

pub fn tracked_users(&self) -> Result<Promise, JsError>

Get the list of users whose devices we are currently tracking.

A user can be marked for tracking using the update_tracked_users method.

Returns a Set<UserId>.

pub fn update_tracked_users(&self, users: &Array) -> Result<Promise, JsError>

Update the list of tracked users.

The OlmMachine maintains a list of users whose devices we are keeping track of: these are known as “tracked users”. These must be users that we share a room with, so that the server sends us updates for their device lists.

Arguments

users - An array of user ids that should be added to the list of tracked users

Any users that hadn’t been seen before will be flagged for a key query immediately, and whenever receive_sync_changes receives a “changed” notification for that user in the future.

Users that were already in the list are unaffected.

pub fn receive_sync_changes( &self, to_device_events: &str, changed_devices: &DeviceLists, one_time_key_counts: &Map, unused_fallback_keys: Option<Set> ) -> Result<Promise, JsError>

Handle to-device events and one-time key counts from a sync response.

This will decrypt and handle to-device events returning the decrypted versions of them.

To decrypt an event from the room timeline call decrypt_room_event.

pub fn outgoing_requests(&self) -> Promise

Get the outgoing requests that need to be sent out.

This returns a list of JsValue to represent either:

KeysUploadRequest,

KeysQueryRequest,

KeysClaimRequest,

ToDeviceRequest,

SignatureUploadRequest,

RoomMessageRequest or

KeysBackupRequest.

Those requests need to be sent out to the server and the responses need to be passed back to the state machine using mark_request_as_sent.

pub fn mark_request_as_sent( &self, request_id: &str, request_type: RequestType, response: &str ) -> Result<Promise, JsError>

Mark the request with the given request ID as sent (see outgoing_requests).

Arguments are:

request_id represents the unique ID of the request that was sent out. This is needed to couple the response with the now sent out request.

response_type represents the type of the request that was sent out.

response represents the response that was received from the server after the outgoing request was sent out.

pub fn encrypt_room_event( &self, room_id: &RoomId, event_type: String, content: &str ) -> Result<Promise, JsError>

Encrypt a room message for the given room.

Beware that a room key needs to be shared before this method can be called using the share_room_key method.

room_id is the ID of the room for which the message should be encrypted. event_type is the type of the event. content is the plaintext content of the message that should be encrypted.

Panics

Panics if a group session for the given room wasn’t shared beforehand.

pub fn decrypt_room_event( &self, event: &str, room_id: &RoomId ) -> Result<Promise, JsError>

Decrypt an event from a room timeline.

Arguments

event, the event that should be decrypted.

room_id, the ID of the room where the event was sent to.

pub fn cross_signing_status(&self) -> Promise

Get the status of the private cross signing keys.

This can be used to check which private cross signing keys we have stored locally.

pub fn export_cross_signing_keys(&self) -> Promise

Export all the private cross signing keys we have.

The export will contain the seeds for the ed25519 keys as unpadded base64 encoded strings.

Returns null if we don’t have any private cross signing keys; otherwise returns a CrossSigningKeyExport.

pub fn import_cross_signing_keys( &self, master_key: Option<String>, self_signing_key: Option<String>, user_signing_key: Option<String> ) -> Promise

Import our private cross signing keys.

The keys should be provided as unpadded-base64-encoded strings.

Returns a CrossSigningStatus.

pub fn bootstrap_cross_signing(&self, reset: bool) -> Promise

Create a new cross signing identity and get the upload request to push the new public keys to the server.

Warning: This will delete any existing cross signing keys that might exist on the server and thus will reset the trust between all the devices.

Uploading these keys will require user interactive auth.

Returns an Array of OutgoingRequests

pub fn get_identity(&self, user_id: &UserId) -> Promise

Get the cross signing user identity of a user.

Returns a promise for an OwnUserIdentity, a UserIdentity, or undefined.

pub fn sign(&self, message: String) -> Promise

Sign the given message using our device key and if available cross-signing master key.

pub fn invalidate_group_session(&self, room_id: &RoomId) -> Promise

Invalidate the currently active outbound group session for the given room.

Returns true if a session was invalidated, false if there was no session to invalidate.

pub fn share_room_key( &self, room_id: &RoomId, users: &Array, encryption_settings: &EncryptionSettings ) -> Result<Promise, JsError>

Get to-device requests to share a room key with users in a room.

room_id is the room ID. users is an array of UserId objects. encryption_settings are an EncryptionSettings object.

Returns an array of ToDeviceRequests.

pub fn get_missing_sessions(&self, users: &Array) -> Result<Promise, JsError>

Get the a key claiming request for the user/device pairs that we are missing Olm sessions for.

Returns null if no key claiming request needs to be sent out, otherwise it returns a KeysClaimRequest object.

Sessions need to be established between devices so group sessions for a room can be shared with them.

This should be called every time a group session needs to be shared as well as between sync calls. After a sync some devices may request room keys without us having a valid Olm session with them, making it impossible to server the room key request, thus it’s necessary to check for missing sessions between sync as well.

Note: Care should be taken that only one such request at a time is in flight, e.g. using a lock.

The response of a successful key claiming requests needs to be passed to the OlmMachine with the mark_request_as_sent.

users represents the list of users that we should check if we lack a session with one of their devices. This can be an empty iterator when calling this method between sync requests.

pub fn get_user_devices(&self, user_id: &UserId) -> Promise

Get a map holding all the devices of a user.

user_id represents the unique ID of the user that the devices belong to.

pub fn get_device(&self, user_id: &UserId, device_id: &DeviceId) -> Promise

Get a specific device of a user if one is found and the crypto store didn’t throw an error.

user_id represents the unique ID of the user that the identity belongs to. device_id represents the unique ID of the device.

pub fn get_verification( &self, user_id: &UserId, flow_id: &str ) -> Result<JsValue, JsError>

Get a verification object for the given user ID with the given flow ID (a to-device request ID if the verification has been requested by a to-device request, or a room event ID if the verification has been requested by a room event).

It returns a “Verification object”, which is either a Sas or Qr object.

pub fn get_verification_request( &self, user_id: &UserId, flow_id: &str ) -> Option<VerificationRequest>

Get a verification request object with the given flow ID.

pub fn get_verification_requests(&self, user_id: &UserId) -> Array

Get all the verification requests of a given user.

pub fn receive_verification_event( &self, event: &str, room_id: &RoomId ) -> Result<Promise, JsError>

Receive a verification event.

This method can be used to pass verification events that are happening in rooms to the OlmMachine. The event should be in the decrypted form.

pub fn export_room_keys(&self, predicate: Function) -> Promise

Export the keys that match the given predicate.

predicate is a closure that will be called for every known InboundGroupSession, which represents a room key. If the closure returns true, the InboundGroupSession will be included in the export, otherwise it won’t.

pub fn import_room_keys( &self, exported_room_keys: &str, progress_listener: Function ) -> Result<Promise, JsError>

Import the given room keys into our store.

exported_keys is a list of previously exported keys that should be imported into our store. If we already have a better version of a key, the key will not be imported.

progress_listener is a closure that takes 2 arguments: progress and total, and returns nothing.

pub fn encrypt_exported_room_keys( exported_room_keys: &str, passphrase: &str, rounds: u32 ) -> Result<String, JsError>

Encrypt the list of exported room keys using the given passphrase.

exported_room_keys is a list of sessions that should be encrypted (it’s generally returned by export_room_keys). passphrase is the passphrase that will be used to encrypt the exported room keys. And rounds is the number of rounds that should be used for the key derivation when the passphrase gets turned into an AES key. More rounds are increasingly computationnally intensive and as such help against brute-force attacks. Should be at least 10_000, while values in the 100_000 ranges should be preferred.

pub fn decrypt_exported_room_keys( encrypted_exported_room_keys: &str, passphrase: &str ) -> Result<String, JsError>

Try to decrypt a reader into a list of exported room keys.

encrypted_exported_room_keys is the result from encrypt_exported_room_keys. passphrase is the passphrase that was used when calling encrypt_exported_room_keys.

pub async fn register_room_key_updated_callback(&self, callback: Function)

Register a callback which will be called whenever there is an update to a room key.

callback should be a function that takes a single argument (an array of {@link RoomKeyInfo}) and returns a Promise.

pub fn close(self)

Shut down the OlmMachine.

The OlmMachine cannot be used after this method has been called.

All associated resources will be closed too, like IndexedDB connections.