Struct matrix_sdk_crypto_js::verification::Sas

pub struct Sas { /* private fields */ }

Short Authentication String (SAS) verification.

Implementations

impl Sas

pub fn user_id(&self) -> UserId

Get our own user ID.

pub fn device_id(&self) -> DeviceId

Get our own device ID.

pub fn other_user_id(&self) -> UserId

Get the user id of the other side.

pub fn other_device_id(&self) -> DeviceId

Get the device ID of the other side.

pub fn flow_id(&self) -> String

Get the unique ID that identifies this SAS verification flow, be either a to-device request ID or a room event ID.

pub fn room_id(&self) -> Option<RoomId>

Get the room ID if the verification is happening inside a room.

pub fn supports_emoji(&self) -> bool

Does this verification flow support displaying emoji for the short authentication string?

pub fn started_from_request(&self) -> bool

Did this verification flow start from a verification request?

pub fn is_self_verification(&self) -> bool

Is this a verification that is verifying one of our own devices?

pub fn have_we_confirmed(&self) -> bool

Have we confirmed that the short auth string matches?

pub fn has_been_accepted(&self) -> bool

Has the verification been accepted by both parties?

pub fn cancel_info(&self) -> Option<CancelInfo>

Get info about the cancellation if the verification flow has been cancelled.

pub fn we_started(&self) -> bool

True if we initiated the verification flow (ie, we sent the m.key.verification.request).

pub fn accept(&self) -> Result<JsValue, JsError>

Accept the SAS verification.

This does nothing (and returns undefined) if the verification was already accepted, otherwise it returns an OutgoingRequest that needs to be sent out.

pub fn confirm(&self) -> Promise

Confirm the SAS verification.

This confirms that the short auth strings match on both sides.

Does nothing if we’re not in a state where we can confirm the short auth string.

Returns a Promise for an array of OutgoingRequests.

pub fn cancel(&self) -> Result<JsValue, JsError>

Cancel the verification.

Returns either an OutgoingRequest which should be sent out, or undefined if the verification is already cancelled.

pub fn cancel_with_code(&self, code: CancelCode) -> Result<JsValue, JsError>

Cancel the verification.

This cancels the verification with given code.

Returns either an OutgoingRequest which should be sent out, or undefined if the verification is already cancelled.

pub fn timed_out(&self) -> bool

Has the SAS verification flow timed out?

pub fn can_be_presented(&self) -> bool

Are we in a state where we can show the short auth string?

pub fn is_done(&self) -> bool

Is the SAS flow done?

pub fn is_cancelled(&self) -> bool

Is the SAS flow cancelled?

pub fn emoji(&self) -> Option<Array>

Get the emoji version of the short auth string.

Returns undefined if we can’t yet present the short auth string, otherwise an array of seven Emoji objects.

pub fn emoji_index(&self) -> Option<Array>

Get the index of the emoji representing the short auth string

Returns undefined if we can’t yet present the short auth string, otherwise seven u8 numbers in the range from 0 to 63 inclusive which can be converted to an emoji using the relevant specification entry.

pub fn decimals(&self) -> Option<Array>

Get the decimal version of the short auth string.

Returns None if we can’t yet present the short auth string, otherwise a tuple containing three 4-digit integers that represent the short auth string.

pub fn register_changes_callback(&self, callback: Function)

Register a callback which will be called whenever there is an update to the request.

The callback is called with no parameters.