Struct matrix_sdk_crypto_js::verification::VerificationRequest

pub struct VerificationRequest { /* private fields */ }

An object controlling key verification requests.

Interactive verification flows usually start with a verification request, this object lets you send and reply to such a verification request.

After the initial handshake the verification flow transitions into one of the verification methods.

Implementations

impl VerificationRequest

pub fn request( own_user_id: &UserId, own_device_id: &DeviceId, other_user_id: &UserId, methods: Option<Array> ) -> Result<String, JsError>

Create an event content that can be sent as a room event to request verification from the other side. This should be used only for verifications of other users and it should be sent to a room we consider to be a DM with the other user.

pub fn own_user_id(&self) -> UserId

Our own user id.

pub fn other_user_id(&self) -> UserId

The ID of the other user that is participating in this verification request.

pub fn other_device_id(&self) -> Option<DeviceId>

The ID of the other device that is participating in this verification.

pub fn room_id(&self) -> Option<RoomId>

Get the room ID if the verification is happening inside a room.

pub fn cancel_info(&self) -> Option<CancelInfo>

Get info about the cancellation if the verification request has been cancelled.

pub fn is_passive(&self) -> bool

Has the verification request been answered by another device?

pub fn is_ready(&self) -> bool

Is the verification request ready to start a verification flow?

pub fn timed_out(&self) -> bool

Has the verification flow timed out?

pub fn time_remaining_millis(&self) -> f64

The number of milliseconds remaining before this verification flow times out.

Returns zero if the time has already passed.

pub fn their_supported_methods(&self) -> Result<Option<Array>, JsError>

Get the supported verification methods of the other side.

Will be present only if the other side requested the verification or if we’re in the ready state.

It return a Option<Vec<VerificationMethod>>.

pub fn our_supported_methods(&self) -> Result<Option<Array>, JsError>

Get our own supported verification methods that we advertised.

Will be present only we requested the verification or if we’re in the ready state.

pub fn flow_id(&self) -> String

Get the unique ID of this verification request.

pub fn is_self_verification(&self) -> bool

Is this a verification that is verifying one of our own devices?

pub fn we_started(&self) -> bool

Did we initiate the verification request?

pub fn is_done(&self) -> bool

Has the verification flow that was started with this request finished?

pub fn phase(&self) -> VerificationRequestPhase

Get the current phase of this request.

Returns a VerificationRequestPhase.

pub fn get_verification(&self) -> JsValue

If this request has transitioned into a concrete verification flow (and not yet been completed or cancelled), returns a Verification object.

Returns: a Sas, a Qr, or undefined.

pub fn register_changes_callback(&self, callback: Function)

Register a callback which will be called whenever there is an update to the request.

The callback is called with no parameters.

pub fn is_cancelled(&self) -> bool

Has the verification flow that was started with this request been cancelled?

pub fn accept_with_methods(&self, methods: Array) -> Result<JsValue, JsError>

Accept the verification request signaling that our client supports the given verification methods.

methods represents the methods that we should advertise as supported by us.

It returns either a ToDeviceRequest, a RoomMessageRequest or undefined.

pub fn accept(&self) -> Result<JsValue, JsError>

Accept the verification request.

This method will accept the request and signal that it supports the m.sas.v1, the m.qr_code.show.v1, and m.reciprocate.v1 method.

m.qr_code.show.v1 will only be signaled if the qrcode feature is enabled. This feature is disabled by default. If it’s enabled and QR code scanning should be supported or QR code showing shouldn’t be supported the accept_with_methods method should be used instead.

It returns either a ToDeviceRequest, a RoomMessageRequest or undefined.

pub fn cancel(&self) -> Result<JsValue, JsError>

Cancel the verification request.

It returns either a ToDeviceRequest, a RoomMessageRequest or undefined.

pub fn start_sas(&self) -> Promise

Transition from this verification request into a SAS verification flow.

Returns Promise<[Sas, RoomMessageRequest|ToDeviceRequest] | undefined>

pub fn generate_qr_code(&self) -> Promise

Available on crate feature qrcode only.

Generate a QR code that can be used by another client to start a QR code based verification.

pub fn scan_qr_code(&self, data: &QrCodeScan) -> Promise

Available on crate feature qrcode only.

Start a QR code verification by providing a scanned QR code for this verification flow.